'Play' icon for run-cell toolbar button
Based on @ellisonbg's research in gh-2914, most tools with similar
functionality to run a piece of code use either the 'play' icon, or
something custom designed (RStudio and Spyder are in the latter camp).

The analogy with audio and video does suggest that 'step forward' is
more similar to running one cell, but it's not clear that users
are actually thinking about it in those terms.

I'm opening this for discussion rather than for immediate merge.